requirements.txt
attrs==19.3.0
Automat==0.8.0
Click==7.0
constantly==15.1.0
cryptography==2.1.4
dominate==2.4.0
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
MarkupSafe==1.1.1
openpyxl==3.0.1
PyHamcrest==1.9.0
PyMySQL==0.9.3
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
Twisted==19.10.0
visitor==0.1.3
Werkzeug==0.16.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.7.1
gevent==20.4.0
